+++ changelog_entry.yaml
@@ -1,0 +1,4 @@
+- bump: patch
+  changes:
+    changed:
+    - Moved to PolicyEngine's L0 package for regularization implementation.
