+++ src/microcalibrate/utils/__init__.py
@@ -1,3 +1,2 @@
-from .l0 import HardConcrete, evaluate_sparse_weights
 from .log_performance import log_performance_over_epochs
 from .metrics import loss, pct_close
